ml/custom_pipeline/requirements.txt
kagglehub==0.3.6
pyyaml==6.0.2
pandas==2.2.3
pillow==11.0.0
scikit-learn==1.5.1
transformers==4.45.2
accelerate==1.2.1
fastapi==0.112.2
uvicorn==0.34.0
matplotlib==3.9.2
seaborn==0.13.2
keras==3.6.0
tensorflow==2.17.0
torch==2.3.1
torchvision==0.18.1
ipykernel==6.29.5